+++ src/array/codec/array_to_bytes/bytes.rs
@@ -120,9 +120,12 @@ fn reverse_endianness(v: &mut [u8], data_type: &DataType) {
 mod tests {
     use std::num::NonZeroU64;
 
-    use crate::array::{
-        codec::{ArrayCodecTraits, CodecTraits},
-        ChunkRepresentation, DataType, FillValue,
+    use crate::{
+        array::{
+            codec::{ArrayCodecTraits, ArrayToBytesCodecTraits, CodecTraits},
+            ChunkRepresentation, ChunkShape, DataType, FillValue,
+        },
+        array_subset::ArraySubset,
     };
 
     use super::*;
@@ -288,5 +291,70 @@ mod tests {
             FillValue::from(num::complex::Complex64::new(0.0, 0.0)),
         )
         .unwrap();
+    }
+
+    #[test]
+    fn codec_bytes_partial_decode() {
+        let chunk_shape: ChunkShape = vec![4, 4].try_into().unwrap();
+        let chunk_representation =
+            ChunkRepresentation::new(chunk_shape.to_vec(), DataType::UInt8, FillValue::from(0u8))
+                .unwrap();
+        let elements: Vec<u8> = (0..chunk_representation.num_elements() as u8).collect();
+        let bytes = elements;
+
+        let codec = BytesCodec::new(None);
+
+        let encoded = codec.encode(bytes, &chunk_representation).unwrap();
+        let decoded_regions = [ArraySubset::new_with_ranges(&[1..3, 0..1])];
+        let input_handle = Box::new(std::io::Cursor::new(encoded));
+        let partial_decoder = codec
+            .partial_decoder(input_handle, &chunk_representation)
+            .unwrap();
+        let decoded_partial_chunk = partial_decoder.partial_decode(&decoded_regions).unwrap();
+
+        let decoded_partial_chunk: Vec<u8> = decoded_partial_chunk
+            .into_iter()
+            .flatten()
+            .collect::<Vec<_>>()
+            .chunks(std::mem::size_of::<u8>())
+            .map(|b| u8::from_ne_bytes(b.try_into().unwrap()))
+            .collect();
+        let answer: Vec<u8> = vec![4, 8];
+        assert_eq!(answer, decoded_partial_chunk);
+    }
+
+    #[cfg(feature = "async")]
+    #[tokio::test]
+    async fn codec_bytes_async_partial_decode() {
+        let chunk_shape: ChunkShape = vec![4, 4].try_into().unwrap();
+        let chunk_representation =
+            ChunkRepresentation::new(chunk_shape.to_vec(), DataType::UInt8, FillValue::from(0u8))
+                .unwrap();
+        let elements: Vec<u8> = (0..chunk_representation.num_elements() as u8).collect();
+        let bytes = elements;
+
+        let codec = BytesCodec::new(None);
+
+        let encoded = codec.encode(bytes, &chunk_representation).unwrap();
+        let decoded_regions = [ArraySubset::new_with_ranges(&[1..3, 0..1])];
+        let input_handle = Box::new(std::io::Cursor::new(encoded));
+        let partial_decoder = codec
+            .async_partial_decoder(input_handle, &chunk_representation)
+            .await
+            .unwrap();
+        let decoded_partial_chunk = partial_decoder
+            .partial_decode(&decoded_regions)
+            .await
+            .unwrap();
+
+        let decoded_partial_chunk: Vec<u8> = decoded_partial_chunk
+            .into_iter()
+            .flatten()
+            .collect::<Vec<_>>()
+            .chunks(std::mem::size_of::<u8>())
+            .map(|b| u8::from_ne_bytes(b.try_into().unwrap()))
+            .collect();
+        let answer: Vec<u8> = vec![4, 8];
+        assert_eq!(answer, decoded_partial_chunk);
     }
 }