+++ src/array/codec/bytes_to_bytes/zstd.rs
@@ -107,4 +107,43 @@ mod tests {
         let answer: Vec<u16> = vec![2, 3, 5];
         assert_eq!(answer, decoded_partial_chunk);
     }
+
+    #[cfg(feature = "async")]
+    #[tokio::test]
+    #[cfg_attr(miri, ignore)]
+    async fn codec_zstd_async_partial_decode() {
+        let elements: Vec<u16> = (0..8).collect();
+        let bytes = crate::array::transmute_to_bytes_vec(elements);
+        let bytes_representation = BytesRepresentation::FixedSize(bytes.len() as u64);
+
+        let configuration: ZstdCodecConfiguration = serde_json::from_str(JSON_VALID).unwrap();
+        let codec = ZstdCodec::new_with_configuration(&configuration);
+
+        let encoded = codec.encode(bytes).unwrap();
+        let decoded_regions = [
+            ByteRange::FromStart(4, Some(4)),
+            ByteRange::FromStart(10, Some(2)),
+        ];
+
+        let input_handle = Box::new(std::io::Cursor::new(encoded));
+        let partial_decoder = codec
+            .async_partial_decoder(input_handle, &bytes_representation)
+            .await
+            .unwrap();
+        let decoded_partial_chunk = partial_decoder
+            .partial_decode(&decoded_regions)
+            .await
+            .unwrap()
+            .unwrap();
+
+        let decoded_partial_chunk: Vec<u16> = decoded_partial_chunk
+            .into_iter()
+            .flatten()
+            .collect::<Vec<_>>()
+            .chunks(std::mem::size_of::<u16>())
+            .map(|b| u16::from_ne_bytes(b.try_into().unwrap()))
+            .collect();
+        let answer: Vec<u16> = vec![2, 3, 5];
+        assert_eq!(answer, decoded_partial_chunk);
+    }
 }